Update Show-Menu.ps1
Pass through the description field of the selected menuitem to the scriptblock. The field is not shown anywhere and therefore can be used to give the scriptblock awareness of what was selected (e.g. a unique reference to the object representing that line in the menu, it could be a user or a folder or a program to run or whatever). The challenge is that since the scriptblock is not invoked until this invoke action is triggered, any variables that are used in the scriptblock are not expanded before this point. The other solution is to call new-menuitem with -action [scriptblock]::create("plaintext representation of my code"), but this has the disadvantage of losing Intellisense for the contents of the scriptblock. I would suggest maybe adding an "ArgumentList" to new-menuitem instead, but this tiny change allows me to proceed with my project.